requirements.txt
grpcio==1.59.0
grpcio-tools==1.59.0
sqlalchemy>=1.4.23
alembic>=1.7.0
protobuf>=4.24.0 